NotificationManager code cleaning.

+++ src/main/java/com/github/ilpersi/BHBot/NotificationManager.java
@@ -57,8 +57,8 @@ public class NotificationManager {
 
        // periodic notification
         if (shouldNotify ()) {
-            if (((Misc.getTime() - timeLastPOAlive) > (bot.settings.poNotifyAlive * Misc.Durations.HOUR)) && timeLastPOAlive != 0 ||
-                    ((Misc.getTime() - timeLastDiscordAlive) > (bot.settings.discordNotifyAlive * Misc.Durations.HOUR)) && timeLastDiscordAlive != 0 ) {
+            if (((Misc.getTime() - timeLastPOAlive) > ((long) bot.settings.poNotifyAlive * Misc.Durations.HOUR)) && timeLastPOAlive != 0 ||
+                    ((Misc.getTime() - timeLastDiscordAlive) > ((long) bot.settings.discordNotifyAlive * Misc.Durations.HOUR)) && timeLastDiscordAlive != 0 ) {
 
                 String aliveScreenName = bot.saveGameScreen("alive-screen");
                 File aliveScreenFile = aliveScreenName != null ? new File(aliveScreenName) : null;
@@ -130,12 +130,12 @@ public class NotificationManager {
                 }
                 aliveMsg.append("\n");
 
-                if ((Misc.getTime() - timeLastPOAlive) > (bot.settings.poNotifyAlive * Misc.Durations.HOUR) && timeLastPOAlive != 0) {
+                if ((Misc.getTime() - timeLastPOAlive) > ((long) bot.settings.poNotifyAlive * Misc.Durations.HOUR) && timeLastPOAlive != 0) {
                     timeLastPOAlive = Misc.getTime();
                     poManager.sendPushOverMessage("Alive notification", aliveMsg.toString(), MessagePriority.QUIET, aliveScreenFile);
                 }
 
-                if ((Misc.getTime() - timeLastDiscordAlive) > (bot.settings.discordNotifyAlive * Misc.Durations.HOUR) && timeLastDiscordAlive != 0) {
+                if ((Misc.getTime() - timeLastDiscordAlive) > ((long) bot.settings.discordNotifyAlive * Misc.Durations.HOUR) && timeLastDiscordAlive != 0) {
                     timeLastDiscordAlive = Misc.getTime();
                     discordManager.sendDiscordMessage(aliveMsg.toString(), aliveScreenFile);
                 }
